Refactored accepting clients.
Previously the code failed when the client was behind a NAT'ed router.
Either personal or at the ISP.
Made connection attempts more solid.
Please note that this solution is not safe in terms of security.
But it is safe in terms of the correctness we strive for.
TODO: assert that inactive clients are removed after a while,
atm they wont be if their tcp connection is still pinging.