Implemented login user story
Note that the authentication itself of the client is secure.
Creating a new account is not as it leaks the hash of the clients account.
Even after logging in communication is not encrypted.
This user story was implemented solely to be able to identify clients.
TODO: properly dispose of CryptoHelper when Account is destroyed.
TODO: properly propagate exceptions to the client.